requirements.txt
Flask==3.1.2
flask-cors==6.0.2
gunicorn==21.2.0

psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
Flask-Migrate==4.0.5

PyJWT==2.8.0
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2
cryptography==41.0.3
Authlib==1.3.1

google-genai==1.62.0
pydantic>=2.9.0
loguru==0.7.2
python-dotenv==1.0.1
resend>=2.29.0

redis==5.0.1
celery==5.6.2